+++ recipes/cask.rb
@@ -20,7 +20,7 @@
 homebrew_tap 'caskroom/cask'
 
 directory '/Library/Caches/Homebrew/Casks' do
-  owner homebrew_owner
+  owner Homebrew.owner
   mode '775'
   only_if { ::Dir.exist?('/Library/Caches/Homebrew') }
 end
