+++ resources/cask.rb
@@ -1,13 +1,11 @@
 property :name, String, regex: /^[\w-]+$/, name_property: true
 property :options, String
 
-include ::Homebrew::Mixin
-
 action :install do
   execute "installing cask #{new_resource.name}" do
     command "/usr/local/bin/brew cask install #{new_resource.name} #{new_resource.options}"
-    user homebrew_owner
-    environment lazy { { 'HOME' => ::Dir.home(homebrew_owner), 'USER' => homebrew_owner } }
+    user Homebrew.owner
+    environment lazy { { 'HOME' => ::Dir.home(Homebrew.owner), 'USER' => Homebrew.owner } }
     not_if { casked? }
   end
 end
@@ -15,8 +13,8 @@ end
 action :uninstall do
   execute "uninstalling cask #{new_resource.name}" do
     command "/usr/local/bin/brew cask uninstall #{new_resource.name}"
-    user homebrew_owner
-    environment lazy { { 'HOME' => ::Dir.home(homebrew_owner), 'USER' => homebrew_owner } }
+    user Homebrew.owner
+    environment lazy { { 'HOME' => ::Dir.home(Homebrew.owner), 'USER' => Homebrew.owner } }
     only_if { casked? }
   end
 end
